refactor: rename 'game-card' to 'start-game-card'
