+++ src/entities/game/ui/index.ts
@@ -1,1 +1,1 @@
-export { default as GameCard } from './game-card.vue';
+export { default as StartGameCard } from './start-game-card.vue';
