+++ src/entities/game/ui/start-game-card.vue
